+++ src/libhidpp/hidpp20/MacroFormat.cpp
@@ -144,8 +144,8 @@ MacroFormat::writeItem (std::vector<uint8_t>::iterator it,
 		jump_addr_it = it;
 		return it+4;
 	case Macro::Item::MousePointer:
-		it = writeBE<int16_t> (it, item.mouseX ());
-		it = writeBE<int16_t> (it, item.mouseY ());
+		it = writeBE<int16_t> (it, static_cast<short> (item.mouseX ()));
+		it = writeBE<int16_t> (it, static_cast<short> (item.mouseY ()));
 		return it;
 	default:
 		return it;
